--- src/boltz_binder_design/esm/__init__.py
@@ -1,1 +1,0 @@
-from .esm import ESM2
